[FEAT] : 유저 프로필 페이지에 배치 API 로 데 연동 KAN-469
